+++ src/imio/migrator/migrator.py
@@ -174,6 +174,16 @@ class Migrator(object):
         portal_factory.manage_setPortalFactoryTypes(listOfTypeIds=registeredFactoryTypes)
         logger.info('Done.')
 
+    def reindexIndexesFor(self, idxs=[], **query):
+        """ Reindex p_idxs on objects of given p_portal_types. """
+        catalog = api.portal.get_tool('portal_catalog')
+        brains = catalog(**query)
+        logger.info('Reindexing indexes "{0}" on "{1}" objects...'.format(', '.join(idxs), len(brains)))
+        for brain in brains:
+            obj = brain.getObject()
+            obj.reindexObject(idxs=idxs)
+        logger.info('Done.')
+
     def install(self, products):
         """ Allows to install a series of products """
         qi = api.portal.get_tool('portal_quickinstaller')